moved transaction structures into transaction.h/.cpp
